added search users by keyword method for autocomplete functionality
