+++ src/main/java/com/cegeka/academy/web/rest/UserResource.java
@@ -7,7 +7,9 @@ import com.cegeka.academy.security.AuthoritiesConstants;
 import com.cegeka.academy.service.MailService;
 import com.cegeka.academy.service.UserService;
 import com.cegeka.academy.service.dto.UserDTO;
-import com.cegeka.academy.web.rest.errors.*;
+import com.cegeka.academy.web.rest.errors.BadRequestAlertException;
+import com.cegeka.academy.web.rest.errors.EmailAlreadyUsedException;
+import com.cegeka.academy.web.rest.errors.LoginAlreadyUsedException;
 import io.github.jhipster.web.util.HeaderUtil;
 import io.github.jhipster.web.util.PaginationUtil;
 import io.github.jhipster.web.util.ResponseUtil;
@@ -186,17 +188,5 @@ public class UserResource {
         log.debug("REST request to delete User: {}", login);
         userService.deleteUser(login);
         return ResponseEntity.noContent().headers(HeaderUtil.createAlert(applicationName,  "A user is deleted with identifier " + login, login)).build();
-    }
-
-    @GetMapping("/{firstName}/{lastName}")
-    public List<UserDTO> getAllUsersByFirstAndLastName(@PathVariable String firstName, @PathVariable String lastName) throws NotFoundException, InvalidArgumentsException {
-
-        return userService.findUsersByFirstAndLastName(firstName, lastName);
-    }
-
-    @GetMapping("/{categoryName}")
-    public List<UserDTO> getAllUsersByCategoryName(@PathVariable String categoryName) throws NotFoundException {
-
-        return userService.findByInterestedCategoryName(categoryName);
     }
 }
